server/requirements.txt
fastapi[standard]
google-api-python-client
python-dotenv
pydantic